gateway/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
aiokafka==0.10.0
python-dotenv==1.0.1
